backend/organization/serializers.py
from rest_framework import serializers
from .models import Business, Branch

class BusinessSerializer(serializers.ModelSerializer):
    class Meta:
        model = Business
        fields = '__all__'

class BranchSerializer(serializers.ModelSerializer):
    business_name = serializers.CharField(source='business.name', read_only=True)

    class Meta:
        model = Branch
        fields = '__all__'
